handle settings and permissions
Need to check that location services are working and that the user has permissions when activity starts.
If location service is not active, then try to resolve the error. if resolution fails, then give the user a rational for needing the service. Allow them to try again or exit.
If permission to use location service is not active, then try to resolve the error. if resolution fails, then give the user a rationale for needing permissions. Allow them to try again or exit. If the user enter 'do not ask again' during resolution. the app will continually show rational till they exit.
Also, including LocationSettingsFailureDialogFragment incase a solution for location service problem cannot be resolved.
Removed LocationFragment and location stuff in MainActivity.