Handle location settings toggle through notification bar
using the notification bar does not pause, resume app so needed a way keep up with changes to location settings happening throught the notification bar. If I were to start over, I would use a Broadcast receiver. However, in this commit is code that uses the locationManager and network provider to check for disabled location settings and then begins the request cycle.